pyproject.toml
[project]
name = "simple-nlp-answering-system"
version = "0.1.0"
description = "Simple NLP answering system with Pinecone embeddings"
readme = "README.md"
requires-python = ">=3.10, <3.13"
dependencies = [
    "fastapi",
    "uvicorn",
    "pinecone",
    "sentence-transformers",
    "fastembed",
    "spacy",
    "en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1.tar.gz",
    "python-dotenv",
    "pydantic",
    "requests",
    "ollama",
    "pyyaml",
    "langchain",
    "langchain-huggingface",
    "onnx",
    "onnxruntime",
    "numpy==1.26.4",

]

[project.optional-dependencies]
dev = [
    "onnxruntime-gpu"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.metadata]
allow-direct-references = true